requirements.txt
fastapi
uvicorn
litellm
python-dotenv
pandas

